# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Benzina API
  description: '

    API dell''app **Benzina**: distributori vicini con il prezzo più basso, media

    nazionale e andamento dei prezzi dei carburanti in Italia.


    I prezzi sono gli open data del MIMIT (Osservaprezzi Carburanti), aggiornati una

    volta al giorno; valutazioni e recensioni vengono da Google Places.


    **Accesso**: tutte le rotte `/v1` richiedono l''header `X-Firebase-AppCheck`

    (token Firebase App Check dell''app) oppure `X-API-Key`. Vedi `docs/API.md`.

    '
  license:
    name: 'Dati prezzi: MIMIT, licenza IODL 2.0'
  version: 0.1.0
paths:
  /health:
    get:
      tags:
      - Servizio
      summary: Stato del servizio
      description: Non richiede autenticazione. `data_date` è il giorno dell'ultimo
        import MIMIT.
      operationId: health_health_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
  /v1/stations/nearby:
    get:
      tags:
      - Distributori
      summary: Distributori vicini, dal più economico
      description: 'Distributori entro `radius_km` che vendono `fuel`, con il prezzo
        dell''ultimo

        aggiornamento MIMIT e la media nazionale per il confronto.'
      operationId: stations_nearby_v1_stations_nearby_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          maximum: 90
          minimum: -90
          description: Latitudine (WGS84).
          examples:
          - 45.4781
          title: Lat
        description: Latitudine (WGS84).
      - name: lng
        in: query
        required: true
        schema:
          type: number
          maximum: 180
          minimum: -180
          description: Longitudine (WGS84).
          examples:
          - 9.227
          title: Lng
        description: Longitudine (WGS84).
      - name: radius_km
        in: query
        required: false
        schema:
          type: number
          maximum: 50
          exclusiveMinimum: 0
          description: Raggio di ricerca in km.
          default: 5
          title: Radius Km
        description: Raggio di ricerca in km.
      - name: fuel
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Fuel'
          description: Carburante.
          default: benzina
        description: Carburante.
      - name: mode
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Mode'
          description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
          default: self
        description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Numero massimo di risultati.
          default: 50
          title: Limit
        description: Numero massimo di risultati.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearbyResponse'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
  /v1/stations:
    get:
      tags:
      - Distributori
      summary: Più distributori per id
      description: Gli id che non esistono (es. distributori chiusi) vengono omessi.
      operationId: stations_list_v1_stations_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: ids
        in: query
        required: true
        schema:
          type: string
          pattern: ^\d+(,\d+){0,49}$
          description: Fino a 50 id separati da virgola (es. i preferiti).
          examples:
          - 3464,3468
          title: Ids
        description: Fino a 50 id separati da virgola (es. i preferiti).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StationList'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
  /v1/stations/{station_id}:
    get:
      tags:
      - Distributori
      summary: Dettaglio di un distributore
      operationId: station_detail_v1_stations__station_id__get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: integer
          description: '`idImpianto` del MIMIT.'
          examples:
          - 3464
          title: Station Id
        description: '`idImpianto` del MIMIT.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Station'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non trovato
  /v1/stations/{station_id}/trend:
    get:
      tags:
      - Andamento prezzi
      summary: Storico prezzi di un distributore
      description: Prezzo a fine giornata. Mancano i giorni prima del primo prezzo
        noto.
      operationId: station_trend_v1_stations__station_id__trend_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: integer
          description: '`idImpianto` del MIMIT.'
          examples:
          - 3464
          title: Station Id
        description: '`idImpianto` del MIMIT.'
      - name: fuel
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Fuel'
          description: Carburante.
          default: benzina
        description: Carburante.
      - name: mode
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Mode'
          description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
          default: self
        description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 366
          minimum: 2
          description: Numero di giorni, oggi incluso.
          default: 30
          title: Days
        description: Numero di giorni, oggi incluso.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non trovato
  /v1/stations/{station_id}/google-rating:
    get:
      tags:
      - Distributori
      summary: Valutazione e recensioni Google
      description: 'Dati di Google Places. Nell''app vanno mostrati con il testo di
        `attribution`

        e con il nome (e link) dell''autore di ogni recensione.


        - 404: distributore inesistente o nessun luogo Google abbinato.

        - 503: integrazione Google non configurata sul server.'
      operationId: station_google_rating_v1_stations__station_id__google_rating_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: integer
          description: '`idImpianto` del MIMIT.'
          examples:
          - 3464
          title: Station Id
        description: '`idImpianto` del MIMIT.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleRating'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non trovato
        '502':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Servizio esterno non disponibile
        '503':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Servizio non disponibile
  /v1/trends/national:
    get:
      tags:
      - Andamento prezzi
      summary: Media nazionale giornaliera
      description: 'Media dei prezzi di tutti i distributori d''Italia, un punto per
        giorno di

        import. Esclusi i prezzi comunicati da oltre 30 giorni e i valori anomali.'
      operationId: trend_national_v1_trends_national_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: fuel
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Fuel'
          description: Carburante.
          default: benzina
        description: Carburante.
      - name: mode
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Mode'
          description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
          default: self
        description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 366
          minimum: 2
          description: Numero di giorni, oggi incluso.
          default: 30
          title: Days
        description: Numero di giorni, oggi incluso.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
  /v1/trends/area:
    get:
      tags:
      - Andamento prezzi
      summary: Media giornaliera della zona
      description: Media dei distributori entro `radius_km`, per confrontarla con
        quella nazionale.
      operationId: trend_area_v1_trends_area_get
      security:
      - AppCheck: []
      - ApiKey: []
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          maximum: 90
          minimum: -90
          description: Latitudine (WGS84).
          examples:
          - 45.4781
          title: Lat
        description: Latitudine (WGS84).
      - name: lng
        in: query
        required: true
        schema:
          type: number
          maximum: 180
          minimum: -180
          description: Longitudine (WGS84).
          examples:
          - 9.227
          title: Lng
        description: Longitudine (WGS84).
      - name: radius_km
        in: query
        required: false
        schema:
          type: number
          maximum: 50
          exclusiveMinimum: 0
          description: Raggio di ricerca in km.
          default: 5
          title: Radius Km
        description: Raggio di ricerca in km.
      - name: fuel
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Fuel'
          description: Carburante.
          default: benzina
        description: Carburante.
      - name: mode
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Mode'
          description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
          default: self
        description: '`self` o `servito`. Ignorato per GPL e metano (si usa `any`).'
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 366
          minimum: 2
          description: Numero di giorni, oggi incluso.
          default: 30
          title: Days
        description: Numero di giorni, oggi incluso.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trend'
        '401':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Non autorizzato
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Parametri non validi
components:
  schemas:
    Fuel:
      type: string
      enum:
      - benzina
      - diesel
      - gpl
      - metano
      title: Fuel
    GoogleRating:
      properties:
        place_id:
          type: string
          title: Place Id
        rating:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating
          examples:
          - 4.3
        rating_count:
          type: integer
          title: Rating Count
          examples:
          - 212
        maps_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Maps Url
          description: Pagina del luogo su Google Maps.
        reviews:
          items:
            $ref: '#/components/schemas/Review'
          type: array
          title: Reviews
          description: Al massimo 5, scelte da Google.
        attribution:
          type: string
          title: Attribution
          description: Testo di attribuzione obbligatorio da mostrare nell'app.
          default: Valutazioni e recensioni fornite da Google
      type: object
      required:
      - place_id
      - rating
      - rating_count
      - maps_url
      - reviews
      title: GoogleRating
    Health:
      properties:
        status:
          type: string
          title: Status
          default: ok
        data_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Data Date
      type: object
      required:
      - data_date
      title: Health
    Mode:
      type: string
      enum:
      - self
      - servito
      - any
      title: Mode
    NearbyResponse:
      properties:
        fuel:
          $ref: '#/components/schemas/Fuel'
        mode:
          $ref: '#/components/schemas/Mode'
          description: 'Modalità effettiva: `any` per GPL e metano.'
        data_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Data Date
          description: Giorno dell'ultimo aggiornamento MIMIT.
        national_average:
          anyOf:
          - type: number
          - type: 'null'
          title: National Average
          description: Media nazionale dello stesso giorno.
        offers:
          items:
            $ref: '#/components/schemas/Offer'
          type: array
          title: Offers
          description: Dal più economico; a parità di prezzo il più vicino.
      type: object
      required:
      - fuel
      - mode
      - data_date
      - national_average
      - offers
      title: NearbyResponse
    Offer:
      properties:
        station:
          $ref: '#/components/schemas/StationSummary'
        price:
          type: number
          title: Price
          examples:
          - 1.739
        mode:
          $ref: '#/components/schemas/Mode'
        reported_at:
          type: string
          format: date-time
          title: Reported At
        distance_km:
          type: number
          title: Distance Km
          examples:
          - 1.2
      type: object
      required:
      - station
      - price
      - mode
      - reported_at
      - distance_km
      title: Offer
    Price:
      properties:
        fuel:
          $ref: '#/components/schemas/Fuel'
        mode:
          $ref: '#/components/schemas/Mode'
          description: '`self` o `servito`.'
        price:
          type: number
          title: Price
          description: €/l (€/kg per il metano).
          examples:
          - 1.739
        reported_at:
          type: string
          format: date-time
          title: Reported At
          description: Quando il gestore ha comunicato il prezzo.
      type: object
      required:
      - fuel
      - mode
      - price
      - reported_at
      title: Price
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
          examples:
          - Non autorizzato
        status:
          type: integer
          title: Status
          examples:
          - 401
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          examples:
          - Token App Check mancante o non valido.
      type: object
      required:
      - title
      - status
      title: Problem
      description: Errore nel formato RFC 9457 (application/problem+json).
    Review:
      properties:
        author:
          type: string
          title: Author
        author_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Author Uri
          description: Profilo Google dell'autore (da mostrare).
        rating:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Rating
        relative_time:
          type: string
          title: Relative Time
          examples:
          - 2 settimane fa
        text:
          type: string
          title: Text
      type: object
      required:
      - author
      - author_uri
      - rating
      - relative_time
      - text
      title: Review
    Station:
      properties:
        id:
          type: integer
          title: Id
          description: '`idImpianto` del MIMIT.'
          examples:
          - 3464
        brand:
          type: string
          title: Brand
          examples:
          - Agip Eni
        name:
          type: string
          title: Name
          examples:
          - 19829 AGRIGENTO
        address:
          type: string
          title: Address
          examples:
          - SS.189 KM. 64+649
        city:
          type: string
          title: City
          examples:
          - AGRIGENTO
        province:
          type: string
          title: Province
          examples:
          - AG
        lat:
          type: number
          title: Lat
          examples:
          - 37.333935
        lng:
          type: number
          title: Lng
          examples:
          - 13.595533
        operator:
          type: string
          title: Operator
          description: Gestore.
        kind:
          type: string
          title: Kind
          description: '`Stradale` o `Autostradale`.'
        prices:
          items:
            $ref: '#/components/schemas/Price'
          type: array
          title: Prices
          description: Prezzi dell'ultimo aggiornamento.
      type: object
      required:
      - id
      - brand
      - name
      - address
      - city
      - province
      - lat
      - lng
      - operator
      - kind
      - prices
      title: Station
    StationList:
      properties:
        stations:
          items:
            $ref: '#/components/schemas/Station'
          type: array
          title: Stations
      type: object
      required:
      - stations
      title: StationList
    StationSummary:
      properties:
        id:
          type: integer
          title: Id
          description: '`idImpianto` del MIMIT.'
          examples:
          - 3464
        brand:
          type: string
          title: Brand
          examples:
          - Agip Eni
        name:
          type: string
          title: Name
          examples:
          - 19829 AGRIGENTO
        address:
          type: string
          title: Address
          examples:
          - SS.189 KM. 64+649
        city:
          type: string
          title: City
          examples:
          - AGRIGENTO
        province:
          type: string
          title: Province
          examples:
          - AG
        lat:
          type: number
          title: Lat
          examples:
          - 37.333935
        lng:
          type: number
          title: Lng
          examples:
          - 13.595533
      type: object
      required:
      - id
      - brand
      - name
      - address
      - city
      - province
      - lat
      - lng
      title: StationSummary
    Trend:
      properties:
        fuel:
          $ref: '#/components/schemas/Fuel'
        mode:
          $ref: '#/components/schemas/Mode'
        points:
          items:
            $ref: '#/components/schemas/TrendPoint'
          type: array
          title: Points
          description: Un punto per giorno, dal più vecchio.
      type: object
      required:
      - fuel
      - mode
      - points
      title: Trend
    TrendPoint:
      properties:
        day:
          type: string
          format: date
          title: Day
        price:
          type: number
          title: Price
      type: object
      required:
      - day
      - price
      title: TrendPoint
  securitySchemes:
    AppCheck:
      type: apiKey
      description: Token Firebase App Check dell'app (produzione).
      in: header
      name: X-Firebase-AppCheck
    ApiKey:
      type: apiKey
      description: Chiave statica per sviluppo e chiamate da server.
      in: header
      name: X-API-Key